new: replaced global conflict list by conflict list per layer, similar to datasets
git-svn-id: http://josm.openstreetmap.de/svn/trunk@1750 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ test/unit/org/openstreetmap/josm/data/osm/history/HistoryWayTest.java
@@ -101,8 +101,8 @@ public class HistoryWayTest {
         }
 
         assertEquals(2, ids.size());
-        assertEquals(1, ids.get(0));
-        assertEquals(2, ids.get(1));
+        assertEquals(1, (long) ids.get(0));
+        assertEquals(2, (long) ids.get(1));
     }
 
 }
